+++ lib/generators/ember/controller_generator.rb
@@ -11,13 +11,13 @@ module Ember
 
       class_option :javascript_engine, :desc => "Engine for JavaScripts"
       class_option :array, :type => :boolean, :default => false, :desc => "Create an Ember.ArrayController to represent multiple objects"
-      class_option :ember_path, :type => :string, :aliases => "-d", :default => false, :desc => "Custom ember app path"
-      class_option :object, :type => :boolean, :default => false, :desc => "Create an Ember.ObjectController to represent a single object"
-      class_option :app_name, :type => :string, :aliases => "-n", :default => false, :desc => "Custom ember app name"
+      class_option :ember_path, :type => :string, :aliases => "-d", :desc => "Custom ember app path"
+      class_option :object, :type => :boolean, :default => false, :desc => "Create an Ember.Controller to represent a single object"
+      class_option :app_name, :type => :string, :aliases => "-n", :desc => "Custom ember app name"
 
 
       def create_controller_files
-        file_path = File.join(ember_path, 'controllers', class_path, "#{file_name}_controller.#{engine_extension}")
+        file_path = File.join(ember_path, 'controllers', class_path, "#{file_name.dasherize}.#{engine_extension}")
 
         if options.array?
           template "array_controller.#{engine_extension}", file_path