+++ lib/generators/ember/template_generator.rb
@@ -9,13 +9,11 @@ module Ember
 
       desc "Creates a new Ember.js template"
       class_option :javascript_engine, :desc => "Engine for JavaScripts"
-      class_option :ember_path, :type => :string, :aliases => "-d", :default => false, :desc => "Custom ember app path"
-
-
+      class_option :ember_path, :type => :string, :aliases => "-d", :desc => "Custom ember app path"
 
       def create_template_files
-        file_path = File.join(ember_path, 'templates', class_path, "#{file_name}.handlebars")
-        template 'template.handlebars', file_path
+        file_path = File.join(ember_path, 'templates', class_path, "#{file_name.dasherize}.hbs")
+        template 'template.hbs', file_path
       end
 
     end