+++ test/generators/bootstrap_generator_test.rb
@@ -18,7 +18,7 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
   def prepare_destination
     super
 
-    copy_directory "app/assets/javascripts"
+    copy_directory "app/assets/javascripts/application.js"
     copy_directory "config"
   end
 
@@ -32,32 +32,57 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
     assert_new_dirs(:skip_git => true)
   end
 
-  %w(js coffee em).each do |engine|
+  test "Assert folder layout with `ember.prefix_dirs`" do
+    begin
+      old, ::Rails.configuration.ember.prefix_dirs = ::Rails.configuration.ember.prefix_dirs, %w(foo)
+
+      run_generator %w(-g)
+
+      %w{foo templates templates/components}.each do |dir|
+        assert_directory ember_path(dir)
+      end
+    ensure
+      ::Rails.configuration.ember.prefix_dirs = old
+    end
+  end
+
+  %w(js coffee em es6).each do |engine|
 
     test "create bootstrap with #{engine} engine" do
+      extension = engine_to_extension(engine)
       run_generator ["--javascript-engine=#{engine}"]
-      assert_file "#{ember_path}/store.js.#{engine}".sub('.js.js','.js')
-      assert_file "#{ember_path}/router.js.#{engine}".sub('.js.js','.js')
-      assert_file "#{ember_path}/#{application_name.underscore}.js.#{engine}".sub('.js.js','.js')
-      #assert_file "#{ember_path}/application.js.#{engine}".sub('.js.js','.js'),
-      #  /Dummy = Ember.Application.create()/
+
+      assert_file ember_path("router.#{extension}")
+      assert_file ember_path("adapters/application.#{extension}")
+      assert_file ember_path("dummy.#{extension}")
+      assert_file ember_path("environment.#{extension}")
+
+      assert_application_file_modified ember_path("application.#{engine}"), 'dummy'
     end
 
     test "create bootstrap with #{engine} engine and custom path" do
       custom_path = ember_path("custom")
+      extension = engine_to_extension(engine)
       run_generator ["--javascript-engine=#{engine}", "-d", custom_path]
-      assert_file "#{custom_path}/store.js.#{engine}".sub('.js.js','.js')
-      assert_file "#{custom_path}/router.js.#{engine}".sub('.js.js','.js')
-      assert_file "#{custom_path}/#{application_name.underscore}.js.#{engine}".sub('.js.js','.js')
-      #assert_file "#{custom_path}/application.js.#{engine}".sub('.js.js','.js'),
-      #  /Dummy = Ember.Application.create()/
+
+      assert_file "#{custom_path}/router.#{extension}"
+      assert_file "#{custom_path}/adapters/application.#{extension}"
+      assert_file "#{custom_path}/dummy.#{extension}"
+      assert_file "#{custom_path}/environment.#{extension}"
+
+      assert_application_file_modified "#{custom_path}/application.#{engine}", 'dummy'
     end
 
     test "create bootstrap with #{engine} and custom app name" do
+      extension = engine_to_extension(engine)
       run_generator ["--javascript-engine=#{engine}", "-n", "MyApp"]
-      assert_file "#{ember_path}/store.js.#{engine}".sub('.js.js','.js'), /MyApp\.Store/
-      assert_file "#{ember_path}/router.js.#{engine}".sub('.js.js','.js'), /MyApp\.Router\.map/
-      assert_file "#{ember_path}/my_app.js.#{engine}".sub('.js.js','.js')
+
+      assert_file ember_path("router.#{extension}"), /MyApp\.Router\.map|Ember\.Router\.extend/
+      assert_file ember_path("adapters/application.#{extension}"), /MyApp\.ApplicationAdapter|DS\.ActiveModelAdapter\.extend/
+      assert_file ember_path("my-app.#{extension}")
+      assert_file ember_path("environment.#{extension}")
+
+      assert_application_file_modified ember_path("application.#{engine}"), 'my-app'
     end
 
   end
@@ -65,7 +90,7 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
   test "adds requires to `application.js`" do
     run_generator
 
-    assert_application_file_modified
+    assert_application_file_modified ember_path('application.js'), 'dummy'
   end
 
   test "creates `application.js` if it doesn't exist" do
@@ -73,7 +98,7 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
 
     run_generator
 
-    assert_application_file_modified
+    assert_application_file_modified ember_path('application.js'), 'dummy'
   end
 
   test "modifies `application.js` it's empty" do
@@ -81,7 +106,7 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
 
     run_generator
 
-    assert_application_file_modified
+    assert_application_file_modified ember_path('application.js'), 'dummy'
   end
 
   test "modifies `application.js` if require_tree doesn't exist and there's no new line" do
@@ -89,7 +114,7 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
 
     run_generator
 
-    assert_application_file_modified
+    assert_application_file_modified ember_path('application.js'), 'dummy'
   end
 
   test "Uses config.ember.app_name as the app name" do
@@ -97,9 +122,8 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
       old, ::Rails.configuration.ember.app_name = ::Rails.configuration.ember.app_name, 'MyApp'
 
       run_generator %w(ember)
-      assert_file "#{ember_path}/store.js", /MyApp\.Store/
-      assert_file "#{ember_path}/store.js", /MyApp\.ApplicationAdapter = DS\.ActiveModelAdapter/
       assert_file "#{ember_path}/router.js", /MyApp\.Router\.map/
+      assert_file "#{ember_path}/adapters/application.js", /MyApp\.ApplicationAdapter = DS\.ActiveModelAdapter/
     ensure
       ::Rails.configuration.ember.app_name = old
     end
@@ -111,7 +135,6 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
       old, ::Rails.configuration.ember.ember_path = ::Rails.configuration.ember.ember_path, custom_path
 
       run_generator
-      assert_file "#{custom_path}/store.js"
       assert_file "#{custom_path}/router.js"
       assert_file "#{custom_path}/#{application_name.underscore}.js"
     ensure
@@ -121,11 +144,10 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
 
   private
 
-  def assert_application_file_modified(application_file = ember_path('application.js'))
-    assert_file application_file, %r{//= require handlebars}
-    assert_file application_file, %r{//= require ember}
-    assert_file application_file, %r{//= require ember-data}
-    assert_file application_file, %r{//= require ./#{application_name.underscore}}
+  def assert_application_file_modified(application_file, application_name)
+    assert_file application_file, %r{(?://|#)= require ember}
+    assert_file application_file, %r{(?://|#)= require ember-data}
+    assert_file application_file, %r{(?://|#)= require \./#{application_name}}
   end
 
   def assert_invoked_generators_files(options = {})
@@ -134,17 +156,16 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
     assert_file "#{path}/application.js"
     assert_file "#{path}/#{application_name}.js"
     assert_file "#{path}/router.js"
-    assert_file "#{path}/store.js"
+    assert_file "#{path}/adapters/application.js"
   end
 
   def assert_new_dirs(options = {})
     path = options[:in_path] || ember_path
 
-    %W{models controllers views helpers components templates templates/components routes}.each do |dir|
+    %W{models controllers views helpers components templates templates/components routes mixins services initializers serializers adapters transforms}.each do |dir|
       assert_directory "#{path}/#{dir}"
       assert_file "#{path}/#{dir}/.gitkeep" unless options[:skip_git]
     end
-
   end
 
   def application_name
@@ -155,5 +176,9 @@ class BootstrapGeneratorTest < Rails::Generators::TestCase
    "app/assets/javascripts/#{custom_path}".chomp('/')
   end
 
+  def engine_to_extension(engine)
+    engine = "module.#{engine}" if engine == 'es6'
+    engine
+  end
 
 end